pyproject.toml
# To install deps from toml
# -------------------------
# > make install

# NOTE: Fix PySide6 xcb plugin error on linux by installing: sudo apt install libxcb-cursor0

[build-system]
requires = [
  "setuptools>=61.0.0",  # for PEP 621 metadata support (fully pyproject.toml projects only)
  "wheel"                # any recent wheel
]
build-backend = "setuptools.build_meta"

[project]
name = "GEMSrun"
description = "Enviroment Runner for GEMS (Graphical Environment Management System)"
version = "2025.4.17.3"
authors = [{name = "Travis L. Seymour, PhD", email = "nogard@ucsc.edu"}]
license = {text = "GPLv3"}
readme = "ReadMe.md"
requires-python = ">=3.11"
keywords = ["VR", "graphical", "environment", "runner"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent"
]

dependencies = [
  "loguru>=0.7.3,<1.0.0",
  "PySide6>=6.9.0,<7.0.0",
  "PyYAML>=6.0.2,<7.0.0",
  "appdirs>=1.4.4,<2.0.0",
  "gTTS>=2.5.4,<3.0.0",
  "pillow>=11.2.1,<12.0.0",
  "six>=1.17.0,<2.0.0",
  "regex>=2024.11.6,<2025.0.0",
  "platformdirs>=4.3.7,<5.0.0",
  "tomli>=2.2.1,<3.0.0",
  "munch>=4.0.0,<5.0.0",
]

[project.urls]
"Bug Tracker" = "https://github.com/travisseymour/GEMSrun/issues"
"Documentation" = "https://github.com/travisseymour/GEMSrun#readme"

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5,<9.0.0",
  "black>=25.1.0,<26.0.0",
  "ruff>=0.11.6,<1.0.0",
  "build>=1.2.2.post1,<2.0.0",
  "plumbum>=1.9.0,<2.0.0",
]

[project.gui-scripts]
gemsrun = "gemsrun.main:main"
GEMSrun = "gemsrun.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["gemsrun*"]

# pull in any files in the resources folder for use with importlib.resources
[tool.setuptools]
include-package-data = true

[tool.black]
line-length = 120

[tool.ruff]
line-length = 120
lint.ignore = ["F821", "F841", "E722"]  # undefined name, assigned but unused variable, bare except